+++ StripeIdentity/StripeIdentity/Source/NativeComponents/Coordinators/VerificationSheetFlowController.swift
@@ -29,6 +29,8 @@ protocol VerificationSheetFlowControllerProtocol: AnyObject {
 
     var documentUploader: DocumentUploaderProtocol? { get }
     var visitedIndividualWelcomePage: Bool { get }
+    var primaryButtonStyle: IdentityVerificationSheet.Configuration.PrimaryButtonStyle { get }
+    var secondaryButtonStyle: IdentityVerificationSheet.Configuration.SecondaryButtonStyle { get }
 
     func transitionToNextScreen(
         skipTestMode: Bool,
@@ -83,6 +85,8 @@ protocol VerificationSheetFlowControllerProtocol: AnyObject {
 final class VerificationSheetFlowController: NSObject {
 
     let brandLogo: UIImage
+    let primaryButtonStyle: IdentityVerificationSheet.Configuration.PrimaryButtonStyle
+    let secondaryButtonStyle: IdentityVerificationSheet.Configuration.SecondaryButtonStyle
     let biometricConsentConfiguration: IdentityVerificationSheet.Configuration.BiometricConsentConfiguration?
 
     weak var delegate: VerificationSheetFlowControllerDelegate?
@@ -97,6 +101,8 @@ final class VerificationSheetFlowController: NSObject {
         configuration: IdentityVerificationSheet.Configuration
     ) {
         self.brandLogo = configuration.brandLogo
+        self.primaryButtonStyle = configuration.primaryButtonStyle
+        self.secondaryButtonStyle = configuration.secondaryButtonStyle
         self.biometricConsentConfiguration = configuration.biometricConsent
     }